# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/signup": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "description": "Render the signup page.",
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "Signup page rendered successfully.",
            "content": {
              "text/html": {}
            }
          }
        }
      },
      "post": {
        "tags": [
          "Authentication"
        ],
        "description": "User signup and auto-login.",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User registration details",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully and logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "User created successfully and logged in."
                }
              }
            }
          },
          "400": {
            "description": "Validation error - Missing required fields or username taken."
          },
          "500": {
            "description": "Internal server error during signup."
          }
        }
      }
    },
    "/login": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "description": "Render the login page.",
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "Login page rendered successfully.",
            "content": {
              "text/html": {}
            }
          }
        }
      },
      "post": {
        "tags": [
          "Authentication"
        ],
        "description": "User login using local authentication strategy.",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User login credentials",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User logged in successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "User logged in successfully."
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials."
          }
        }
      }
    },
...
    "/admin": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Retrieves all necessary data for the admin panel. Admin access only.",
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "Admin panel rendered successfully.",
            "content": {
              "text/html": {}
            }
          },
          "403": {
            "description": "Forbidden - Admin access required.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "Forbidden - Admin access required."
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Failed to load admin panel.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "Failed to load the admin panel."
                }
              }
            }
          }
        }
      }
    },
...
    "/start": {
      "get": {
        "tags": [
          "Start"
        ],
        "description": "Fetch the welcome page with user details or best-rated hotel.",
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved welcome page.",
            "content": {
              "text/html": {
                "schema": {
                  "title": "Welcome",
                  "user": {
                    "id": 1,
                    "username": "john_doe",
                    "role": "User"
                  },
                  "bestHotel": {
                    "id": 5,
                    "name": "Luxury Inn",
                    "avgRating": 4.8
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - Failed to load start page.",
            "content": {
              "text/html": {
                "schema": {
                  "title": "Error",
                  "message": "Failed to load start page."
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "username": {
          "type": "string",
          "example": "exampleuser"
        },
        "firstName": {
          "type": "string",
          "example": "First"
        },
        "lastName": {
          "type": "string",
          "example": "Last"
        },
        "email": {
          "type": "string",
          "example": "user@example.com"
        },
        "role": {
          "type": "string",
          "example": "User"
        }
      },
      "required": [
        "id",
        "username",
        "firstName",
        "lastName",
        "email",
        "role"
      ]
    },
...
  }
}